Update Huntorr.py
Thanks for creating this tool!

If the user is left blank, something that Transmission allows, this wouldn’t work since it sets certain default values. 

A solution is to set the variables in Docker Compose to be blank, as this works.

If this isn't useful, feel free to ignore this PR.

Again, thanks for your work!